#362 test: QueryDslTest 어노테이션 적용
