+++ src/test/java/com/floney/floney/user/repository/SignoutReasonRepositoryTest.java
@@ -1,23 +1,18 @@
 package com.floney.floney.user.repository;
 
-import com.floney.floney.config.TestConfig;
+import com.floney.floney.config.QueryDslTest;
 import com.floney.floney.user.dto.constant.SignoutType;
 import com.floney.floney.user.entity.SignoutReason;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.context.annotation.Import;
 
 import javax.persistence.EntityManager;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DataJpaTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-@Import(TestConfig.class)
+@QueryDslTest
 public class SignoutReasonRepositoryTest {
 
     @Autowired
