+++ src/test/java/com/floney/floney/user/repository/UserRepositoryTest.java
@@ -1,23 +1,17 @@
 package com.floney.floney.user.repository;
 
-import com.floney.floney.config.TestConfig;
+import com.floney.floney.config.QueryDslTest;
 import com.floney.floney.fixture.UserFixture;
 import com.floney.floney.user.entity.User;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.context.annotation.Import;
 
 import static com.floney.floney.common.constant.Status.ACTIVE;
 import static com.floney.floney.fixture.BookFixture.EMAIL;
 
-@DataJpaTest
-@Import(TestConfig.class)
-@AutoConfigureTestDatabase(replace = Replace.NONE)
+@QueryDslTest
 public class UserRepositoryTest {
 
     @Autowired
